+++ src/views/home/index.vue
@@ -1,0 +1,22 @@
+<script setup lang="ts"></script>
+
+<template>
+	<div class="home card">
+		<img class="home-bg" src="@/assets/images/welcome.png" alt="welcome" />
+	</div>
+</template>
+
+<style lang="scss" scoped>
+.home {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	width: 100%;
+	height: 100%;
+
+	.home-bg {
+		width: 70%;
+		margin-bottom: 20px;
+	}
+}
+</style>